added tests for components
